Rename & Relocate activity editor
