+++ src/app/worklists/editor/activity-editor.component.scss
