+++ src/app/worklists/editor/tabs/activity-close.component.scss
@@ -1,0 +1,1 @@
+@import "../../../../styles/form";
