+++ src/app/worklists/editor/tabs/activity-update.component.scss
@@ -1,5 +1,5 @@
-@import "../../../styles/editor-window-style";
-@import "../../../styles/form";
+@import "../../../../styles/editor-window-style";
+@import "../../../../styles/form";
 .form-frame {
   height: 430px;
   overflow-y: auto;
