--- src/app/worklists/editor/task-close.component.scss
@@ -1,1 +1,0 @@
-@import "../../../styles/form";
