+++ src/app/worklists/worklists.module.ts
@@ -16,9 +16,9 @@ import { WorklistsMainPageComponent } from './task-list/worklists-main-page.comp
 
 import { WorklistsRoutingModule } from './worklists-routing.module';
 
-import { TaskFormEditorComponent } from './editor/task-form-editor.component';
-import { TaskCloseComponent } from './editor/task-close.component';
-import { TaskUpdateComponent } from './editor/task-update.component';
+import { ActivityEditorComponent } from './editor/activity-editor.component';
+import { ActivityCloseComponent } from './editor/tabs/activity-close.component';
+import { ActivityUpdateComponent } from './editor/tabs/activity-update.component';
 
 import { ProjectsMainPageComponent } from './main-page/projects-main-page.component';
 import { GanttComponent } from './gantt/gantt.component';
@@ -29,13 +29,6 @@ import { ActivityDescriptionComponent } from './create-activity/activity-descrip
 import { ActivityWorklistComponent } from './create-activity/activity-worklist.component';
 import { ActivityInfoComponent } from './create-activity/activity-info.component';
 
-//import { ActivityEditorComponent } from './create-activity/activity-editor.component';
-//import { ActivityGeneralInfoComponent } from './create-activity/activity-general-info.component';
-//import { ActivityCurrentStateComponent } from './create-activity/activity-current-state.component';
-//import { ActivityNextStateComponent } from './create-activity/activity-next-state.component';
-//import { ActivityControlAndStateComponent } from './create-activity/activity-control-and-state.component';
-//import { ActivityAssignTaskComponent } from './create-activity/activity-assing-task.component';
-
 /**
  * Do not specify providers for modules that might be imported by a lazy loaded module.
  */
@@ -43,11 +36,11 @@ import { ActivityInfoComponent } from './create-activity/activity-info.component
 @NgModule({
   imports: [WorklistsRoutingModule, SharedModule, CommonModule, FormsModule, ContractsModule],
   declarations: [WorklistsMainPageComponent, 
-                 TaskFormEditorComponent, TaskCloseComponent,
+                 ActivityEditorComponent, ActivityCloseComponent,
                  ProjectsMainPageComponent, GanttComponent,
                  ActivityAddComponent,ActivitySelectorComponent, ActivityDescriptionComponent,
                  ActivityWorklistComponent, ActivityInfoComponent,
-                 TaskUpdateComponent],                 
+                 ActivityUpdateComponent],                 
   exports: [WorklistsMainPageComponent,
             ProjectsMainPageComponent, GanttComponent,]
 })
